refactor(dto): move ProcessingMethod DTO to root and add Author field
