+++ Source/DakLakCoffeeSupplyChainBE/DakLakCoffeeSupplyChain.Common/DTOs/ProcessingMethodDTOs/ProcessingMethodDetailDto.cs
@@ -4,13 +4,14 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace DakLakCoffeeSupplyChain.Common.DTOs.Flow3DTOs.ProcessingMethod
+namespace DakLakCoffeeSupplyChain.Common.DTOs.ProcessingMethodDTOs
 {
-    public class ProcessingMethodViewAllDto
+    public class ProcessingMethodDetailDto
     {
         public int MethodId { get; set; }
         public string MethodCode { get; set; } = string.Empty;
         public string Name { get; set; } = string.Empty;
-        public string? Description { get; set; } = string.Empty;
+        public string Description { get; set; } = string.Empty;
+        public int StageCount { get; set; }
     }
 }
