+++ Source/DakLakCoffeeSupplyChainBE/DakLakCoffeeSupplyChain.Common/DTOs/ProcessingMethodDTOs/ProcessingMethodViewAllDto.cs
@@ -4,14 +4,13 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace DakLakCoffeeSupplyChain.Common.DTOs.Flow3DTOs.ProcessingMethod
+namespace DakLakCoffeeSupplyChain.Common.DTOs.ProcessingMethodDTOs
 {
-    public class ProcessingMethodDetailDto
+    public class ProcessingMethodViewAllDto
     {
         public int MethodId { get; set; }
         public string MethodCode { get; set; } = string.Empty;
         public string Name { get; set; } = string.Empty;
-        public string Description { get; set; } = string.Empty;
-        public int StageCount { get; set; }
+        public string? Description { get; set; } = string.Empty;
     }
 }
